pyproject.toml
[project]
name = "bank-support"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.68.1",
    "asyncio>=4.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.117.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "jupyter-core>=5.8.1",
    "logfire>=4.10.0",
    "modal-client>=1.0.0",
    "openai>=1.109.1",
    "opentelemetry-instrumentation-botocore>=0.58b0",
    "opentelemetry-instrumentation-jinja2>=0.58b0",
    "opentelemetry-instrumentation-requests>=0.58b0",
    "opentelemetry-instrumentation-tornado>=0.58b0",
    "opentelemetry-instrumentation-urllib>=0.58b0",
    "pydantic>=2.11.9",
    "pydantic-ai[examples]>=1.0.10",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "uvicorn>=0.37.0",
]
